requirement.txt
Django==2.2.4
django-cors-headers==3.1.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
Pillow==6.2.0
Pygments==2.4.2
PyJWT==1.7.1
pytz==2019.2
sqlparse==0.3.0
